libprjxray: move definition of icap_ecc to source file
